Moving user endpoints to user module
