+++ data/server.js
@@ -10,6 +10,7 @@ const db = new sqlite3.Database('pushpin.db');
 // Our modules
 const init = require('./init');
 const tokens = require('./tokens');
+const userModule = require('./users');
 
 const keyFile = '../key.pem';
 const certFile = '../cert.pem';
@@ -87,67 +88,7 @@ init.initialiseDatabase(db, async () => {
     });
 
     // User Login
-    app.post('/login', async (req, res) => {
-        const email = req.body.email;
-        const password = req.body.password;
-
-        const sql = `
-    SELECT id, firstName, lastName, displayName, imageURL, email, password
-    FROM users
-    WHERE email = ?
-    `
-
-        const stmt = db.prepare(sql);
-
-        stmt.get(email, (err, user) => {
-            // If there was a query error - handle it.
-            if (err) {
-                res.status(500).json({
-                    'error': err.toString()
-                });
-                return;
-            }
-
-            // If no matching user was found, return a login failure message.
-            if (!user) {
-                res.status(400).json({
-                    'error': 'Login failed'
-                });
-                return;
-            }
-
-            bcrypt.compare(password, user.password, async (err, success) => {
-                if (err) {
-                    res.status(500).json({
-                        'error': err.toString()
-                    });
-                }
-
-                if (success) {
-                    delete user.password;
-
-                    const userIpAddress = req.headers['x-forwarded-for'] || req.connection.remoteAddress;
-
-                    try {
-                        const token = await tokens.createToken(db, user.id, req.headers['user-agent'], userIpAddress);
-
-                        res.status(200).json({
-                            "tokenInfo": token,
-                            user
-                        });
-                    } catch (error) {
-                        res.status(500).json({
-                            'error': error.toString()
-                        });
-                    }
-                } else {
-                    res.status(400).json({
-                        'error': "Login failed"
-                    });
-                }
-            });
-        });
-    });
+    app.post('/login', userModule.user.login(db));
 
     // User Logout
     app.get('/logout', tokens.checkTokenMiddleware({"db": db, "debug": debugMode}), async (req, res) => {
